refactor: 경매 상세페이지 UI 이동 완료 (#99)
