+++ apps/web/src/__components/common/ui/PageHeader.tsx
@@ -1,4 +1,4 @@
-import GoBackButton from '../GoBackButton';
+import GoBackButton from '../../../shared/ui/GoBackButton';
 import PageTitle from './PageTitle';
 
 interface PageHeaderProps {
