+++ apps/web/src/features/auction/AuctionDetailICarousel.tsx
