+++ apps/web/src/features/auction/EndingSoonCarousel.tsx
@@ -2,8 +2,8 @@
 import { Carousel, CarouselContent, CarouselItem } from '@repo/ui/components/ui/carousel';
 import NotAuctionImage from 'assets/images/auctionDefault.png';
 
-import AuctionCard from '../common/AuctionCard';
 import { SortedAuctionItemType } from 'src/entities/auction/types';
+import AuctionCard from 'src/features/auction/shared/AuctionCard';
 
 const EndingSoonCarousel = ({ endingSoonAuctions }: { endingSoonAuctions: SortedAuctionItemType[] }) => {
   return (
