+++ apps/web/src/features/auction/EndingSoonListSection.tsx
