+++ apps/web/src/features/auction/SellerInfoSection.tsx
@@ -1,8 +1,8 @@
 import { Card } from '@repo/ui/components/ui/card';
 import { UserInfoType } from 'src/app/api/auth/user-info/route';
-import UserAvatar from 'src/components/common/UserAvatar';
 import { fetchSellerAuctionCount } from 'src/entities/auction/api';
 import { SellerInfoType } from 'src/entities/auction/types';
+import UserAvatar from 'src/shared/ui/UserAvatar';
 
 type SellerInfoSectionProps = SellerInfoType & { userInfo: UserInfoType } & { address: string[] };
 
@@ -13,22 +13,22 @@ const SellerInfoSection = async ({ seller, address, userInfo }: SellerInfoSectio
 
   return (
     <Card className="mb-4 p-5 shadow-sm">
-      <h3 className="font-medium text-(--color-text-base)">판매자 정보</h3>
+      <h3 className="text-(--color-text-base) font-medium">판매자 정보</h3>
       <div className="mb-4 flex items-center">
         <UserAvatar src={seller.avatar!} alt={userNickname!} size="sm" />
         <div>
-          <h3 className="font-sm text-sm text-(--color-text-base)">{userNickname}</h3>
-          <p className="text-sm text-(--color-warm-gray)">{address[0]}</p>
+          <h3 className="font-sm text-(--color-text-base) text-sm">{userNickname}</h3>
+          <p className="text-(--color-warm-gray) text-sm">{address[0]}</p>
         </div>
       </div>
-      <div className="flex text-sm text-(--color-warm-gray)">
+      <div className="text-(--color-warm-gray) flex text-sm">
         <div className="flex-1">
           <p>총 경매</p>
-          <p className="font-medium text-(--color-text-base)">{totalAuctions}개</p>
+          <p className="text-(--color-text-base) font-medium">{totalAuctions}개</p>
         </div>
         <div className="flex-1">
           <p>진행중인 경매</p>
-          <p className="font-medium text-(--color-text-base)">{activeAuctions}개</p>
+          <p className="text-(--color-text-base) font-medium">{activeAuctions}개</p>
         </div>
       </div>
     </Card>
