+++ apps/web/src/shared/ui/AuctionErrorBoundary.tsx
