+++ apps/web/src/shared/ui/UserAvatar.tsx
