+++ apps/web/src/widgets/MainBanner.tsx
